server/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
fastapi==0.104.1
h11==0.16.0
httptools==0.7.1
idna==3.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.8.0
PyMySQL==1.1.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn[standard]==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==12.0